refactor(mcp): split relay internals by ownership
