+++ crates/agentgateway/src/mcp/router.rs
@@ -13,7 +13,7 @@ use crate::http::authorization::RuleSets;
 use crate::http::sessionpersistence::{Encoder, MCPSnapshotMember};
 use crate::http::*;
 use crate::mcp::auth;
-use crate::mcp::handler::RelayInputs;
+use crate::mcp::relay::RelayInputs;
 use crate::mcp::session::SessionManager;
 use crate::mcp::sse::LegacySSEService;
 use crate::mcp::streamablehttp::{StreamableHttpServerConfig, StreamableHttpService};
